fix a lot

+++ ex2.h
@@ -1,13 +1,13 @@
 #include <stdint.h>
 
 #define SOUND_DECLARE(name) \
-extern uint32_t _binary_##name##_raw_start[]; \
-extern uint32_t _binary_##name##_raw_end[];
+extern uint16_t _binary_##name##_raw_start[]; \
+extern uint16_t _binary_##name##_raw_end[];
 #define SOUND(name) {_binary_##name##_raw_start, _binary_##name##_raw_end}
 typedef struct {
-	const uint32_t *const start;
-	const uint32_t *const end;
-	const volatile uint32_t * pos;
+	const uint16_t *const start;
+	const uint16_t *const end;
+	const volatile uint16_t * pos;
 } Sound;
 extern volatile Sound sounds[8];
